# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetControl API — veterinary clinic management system
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  clientes: number;
  pacientes: number;
  consultas: number;
  consultasHoy: number;
  saldo: number;
}

export interface ConsultaReciente {
  id: number;
  fecha: string;
  pacienteId: number;
  paciente: string;
  propietario: string;
  motivo: string;
}

export interface DashboardData {
  stats: DashboardStats;
  recientes: ConsultaReciente[];
}

export interface Cliente {
  id: number;
  nombre: string;
  /** @nullable */
  apellidos?: string | null;
  telefono: string;
  /** @nullable */
  telefonoAlterno?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  rfc?: string | null;
  /** @nullable */
  direccion?: string | null;
  /** @nullable */
  notas?: string | null;
  creadoEn: string;
}

export interface ClienteListItem {
  id: number;
  nombre: string;
  /** @nullable */
  apellidos?: string | null;
  telefono: string;
  /** @nullable */
  email?: string | null;
  totalPacientes: number;
  saldo: number;
}

export interface PacienteListItem {
  id: number;
  nombre: string;
  especie: string;
  /** @nullable */
  raza?: string | null;
  propietario: string;
  clienteId: number;
  /** @nullable */
  peso?: number | null;
  estado: string;
  creadoEn: string;
}

export interface Movimiento {
  id: number;
  clienteId: number;
  /** @nullable */
  pacienteId?: number | null;
  /** @nullable */
  paciente?: string | null;
  fecha: string;
  tipo: string;
  concepto: string;
  importe: number;
  /** @nullable */
  metodoPago?: string | null;
  /** @nullable */
  referencia?: string | null;
  /** @nullable */
  notas?: string | null;
  creadoEn: string;
}

export interface ClienteDetalle {
  cliente: Cliente;
  saldo: number;
  pacientes: PacienteListItem[];
  movimientos: Movimiento[];
}

export interface ClienteInput {
  /** @minLength 1 */
  nombre: string;
  apellidos?: string;
  /** @minLength 1 */
  telefono: string;
  telefonoAlterno?: string;
  email?: string;
  rfc?: string;
  direccion?: string;
  notas?: string;
}

export interface Paciente {
  id: number;
  clienteId: number;
  nombre: string;
  especie: string;
  /** @nullable */
  raza?: string | null;
  /** @nullable */
  sexo?: string | null;
  /** @nullable */
  fechaNacimiento?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  peso?: number | null;
  /** @nullable */
  microchip?: string | null;
  esterilizado: boolean;
  estado: string;
  /** @nullable */
  alergias?: string | null;
  /** @nullable */
  antecedentes?: string | null;
  /** @nullable */
  vacunas?: string | null;
  /** @nullable */
  alimentacion?: string | null;
  creadoEn: string;
}

export interface Consulta {
  id: number;
  pacienteId: number;
  fecha: string;
  /** @nullable */
  hora?: string | null;
  /** @nullable */
  medico?: string | null;
  motivo: string;
  /** @nullable */
  anamnesis?: string | null;
  /** @nullable */
  exploracionFisica?: string | null;
  /** @nullable */
  peso?: number | null;
  /** @nullable */
  temperatura?: number | null;
  /** @nullable */
  frecuenciaCardiaca?: number | null;
  /** @nullable */
  frecuenciaRespiratoria?: number | null;
  /** @nullable */
  mucosas?: string | null;
  /** @nullable */
  trc?: string | null;
  /** @nullable */
  condicionCorporal?: string | null;
  /** @nullable */
  diagnostico?: string | null;
  /** @nullable */
  diagnosticosDiferenciales?: string | null;
  /** @nullable */
  plan?: string | null;
  /** @nullable */
  tratamiento?: string | null;
  /** @nullable */
  pronostico?: string | null;
  /** @nullable */
  proximaCita?: string | null;
  /** @nullable */
  observaciones?: string | null;
  creadoEn: string;
}

export interface RecetaPartida {
  id: number;
  recetaId: number;
  medicamento: string;
  /** @nullable */
  presentacion?: string | null;
  dosis: string;
  /** @nullable */
  via?: string | null;
  /** @nullable */
  frecuencia?: string | null;
  /** @nullable */
  duracion?: string | null;
  /** @nullable */
  instrucciones?: string | null;
}

export interface RecetaDetalle {
  id: number;
  pacienteId: number;
  /** @nullable */
  consultaId?: number | null;
  fecha: string;
  /** @nullable */
  medico?: string | null;
  /** @nullable */
  indicacionesGenerales?: string | null;
  /** @nullable */
  proximaRevision?: string | null;
  partidas: RecetaPartida[];
  paciente: string;
  propietario: string;
  /** @nullable */
  especie?: string | null;
  /** @nullable */
  raza?: string | null;
  creadoEn: string;
}

export interface Prueba {
  id: number;
  pacienteId: number;
  /** @nullable */
  consultaId?: number | null;
  fecha: string;
  /** @nullable */
  tipo?: string | null;
  nombre: string;
  /** @nullable */
  laboratorio?: string | null;
  /** @nullable */
  resultado?: string | null;
  /** @nullable */
  interpretacion?: string | null;
  costo: number;
  creadoEn: string;
}

export interface PacienteDetalle {
  paciente: Paciente;
  propietario: string;
  propietarioId: number;
  consultas: Consulta[];
  recetas: RecetaDetalle[];
  pruebas: Prueba[];
}

export interface PacienteInput {
  clienteId: number;
  /** @minLength 1 */
  nombre: string;
  /** @minLength 1 */
  especie: string;
  raza?: string;
  sexo?: string;
  fechaNacimiento?: string;
  color?: string;
  peso?: number;
  microchip?: string;
  esterilizado?: boolean;
  estado?: string;
  alergias?: string;
  antecedentes?: string;
  vacunas?: string;
  alimentacion?: string;
}

export interface ConsultaInput {
  /** @minLength 1 */
  fecha: string;
  hora?: string;
  medico?: string;
  /** @minLength 1 */
  motivo: string;
  anamnesis?: string;
  exploracionFisica?: string;
  peso?: number;
  temperatura?: number;
  frecuenciaCardiaca?: number;
  frecuenciaRespiratoria?: number;
  mucosas?: string;
  trc?: string;
  condicionCorporal?: string;
  diagnostico?: string;
  diagnosticosDiferenciales?: string;
  plan?: string;
  tratamiento?: string;
  pronostico?: string;
  proximaCita?: string;
  observaciones?: string;
}

export interface RecetaPartidaInput {
  /** @minLength 1 */
  medicamento: string;
  presentacion?: string;
  /** @minLength 1 */
  dosis: string;
  via?: string;
  frecuencia?: string;
  duracion?: string;
  instrucciones?: string;
}

export interface RecetaInput {
  consultaId?: number;
  /** @minLength 1 */
  fecha: string;
  medico?: string;
  indicacionesGenerales?: string;
  proximaRevision?: string;
  /** @minItems 1 */
  partidas: RecetaPartidaInput[];
}

export interface PruebaInput {
  consultaId?: number;
  /** @minLength 1 */
  fecha: string;
  tipo?: string;
  /** @minLength 1 */
  nombre: string;
  laboratorio?: string;
  resultado?: string;
  interpretacion?: string;
  costo?: number;
}

export type MovimientoInputTipo = typeof MovimientoInputTipo[keyof typeof MovimientoInputTipo];


export const MovimientoInputTipo = {
  Cargo: 'Cargo',
  Pago: 'Pago',
  Ajuste: 'Ajuste',
} as const;

export interface MovimientoInput {
  pacienteId?: number;
  /** @minLength 1 */
  fecha: string;
  tipo: MovimientoInputTipo;
  /** @minLength 1 */
  concepto: string;
  /** @minimum 0 */
  importe: number;
  metodoPago?: string;
  referencia?: string;
  notas?: string;
}

export type GetClientesParams = {
q?: string;
};

export type GetPacientesParams = {
q?: string;
};