# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VetControl API — veterinary clinic management system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clientes:
    get:
      operationId: getClientes
      tags: [clientes]
      summary: List clients with optional search
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClienteListItem"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{clienteId}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Get client detail with pets and movements
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteDetalle"
        "404":
          description: Not found
    put:
      operationId: updateCliente
      tags: [clientes]
      summary: Update a client
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
        "404":
          description: Not found
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Delete a client and all associated data
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /clientes/{clienteId}/movimientos:
    post:
      operationId: createMovimiento
      tags: [movimientos]
      summary: Create a financial movement for a client
      parameters:
        - name: clienteId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovimientoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movimiento"
        "404":
          description: Not found
  /movimientos/{movimientoId}:
    delete:
      operationId: deleteMovimiento
      tags: [movimientos]
      summary: Delete a movement
      parameters:
        - name: movimientoId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /pacientes/{pacienteId}/consultas:
    post:
      operationId: createConsulta
      tags: [consultas]
      summary: Create a consultation for a patient
      parameters:
        - name: pacienteId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consulta"
        "404":
          description: Not found
  /consultas/{consultaId}:
    get:
      operationId: getConsulta
      tags: [consultas]
      summary: Get a consultation
      parameters:
        - name: consultaId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consulta"
        "404":
          description: Not found
    put:
      operationId: updateConsulta
      tags: [consultas]
      summary: Update a consultation
      parameters:
        - name: consultaId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConsultaInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consulta"
        "404":
          description: Not found
    delete:
      operationId: deleteConsulta
      tags: [consultas]
      summary: Delete a consultation
      parameters:
        - name: consultaId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Cliente:
      type: object
      properties:
        id:
          type: number
        nombre:
          type: string
        apellidos:
          type: ["string", "null"]
        telefono:
          type: string
        telefonoAlterno:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        rfc:
          type: ["string", "null"]
        direccion:
          type: ["string", "null"]
        notas:
          type: ["string", "null"]
        creadoEn:
          type: string
      required:
        - id
        - nombre
        - telefono
        - creadoEn
    ClienteListItem:
      type: object
      properties:
        id:
          type: number
        nombre:
          type: string
        apellidos:
          type: ["string", "null"]
        telefono:
          type: string
        email:
          type: ["string", "null"]
        totalPacientes:
          type: number
        saldo:
          type: number
      required:
        - id
        - nombre
        - telefono
        - totalPacientes
        - saldo
    ClienteDetalle:
      type: object
      properties:
        cliente:
          $ref: "#/components/schemas/Cliente"
        saldo:
          type: number
        pacientes:
          type: array
          items:
            $ref: "#/components/schemas/PacienteListItem"
        movimientos:
          type: array
          items:
            $ref: "#/components/schemas/Movimiento"
      required:
        - cliente
        - saldo
        - pacientes
        - movimientos
    ClienteInput:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
        apellidos:
          type: string
        telefono:
          type: string
          minLength: 1
        telefonoAlterno:
          type: string
        email:
          type: string
        rfc:
          type: string
        direccion:
          type: string
        notas:
          type: string
      required:
        - nombre
        - telefono
...
    PacienteListItem:
      type: object
      properties:
        id:
          type: number
        nombre:
          type: string
        especie:
          type: string
        raza:
          type: ["string", "null"]
        propietario:
          type: string
        clienteId:
          type: number
        peso:
          type: ["number", "null"]
        estado:
          type: string
        creadoEn:
          type: string
      required:
        - id
        - nombre
        - especie
        - propietario
        - clienteId
        - estado
        - creadoEn
...
    Consulta:
      type: object
      properties:
        id:
          type: number
        pacienteId:
          type: number
        fecha:
          type: string
        hora:
          type: ["string", "null"]
        medico:
          type: ["string", "null"]
        motivo:
          type: string
        anamnesis:
          type: ["string", "null"]
        exploracionFisica:
          type: ["string", "null"]
        peso:
          type: ["number", "null"]
        temperatura:
          type: ["number", "null"]
        frecuenciaCardiaca:
          type: ["number", "null"]
        frecuenciaRespiratoria:
          type: ["number", "null"]
        mucosas:
          type: ["string", "null"]
        trc:
          type: ["string", "null"]
        condicionCorporal:
          type: ["string", "null"]
        diagnostico:
          type: ["string", "null"]
        diagnosticosDiferenciales:
          type: ["string", "null"]
        plan:
          type: ["string", "null"]
        tratamiento:
          type: ["string", "null"]
        pronostico:
          type: ["string", "null"]
        proximaCita:
          type: ["string", "null"]
        observaciones:
          type: ["string", "null"]
        creadoEn:
          type: string
      required:
        - id
        - pacienteId
        - fecha
        - motivo
        - creadoEn
    ConsultaInput:
      type: object
      properties:
        fecha:
          type: string
          minLength: 1
        hora:
          type: string
        medico:
          type: string
        motivo:
          type: string
          minLength: 1
        anamnesis:
          type: string
        exploracionFisica:
          type: string
        peso:
          type: number
        temperatura:
          type: number
        frecuenciaCardiaca:
          type: number
        frecuenciaRespiratoria:
          type: number
        mucosas:
          type: string
        trc:
          type: string
        condicionCorporal:
          type: string
        diagnostico:
          type: string
        diagnosticosDiferenciales:
          type: string
        plan:
          type: string
        tratamiento:
          type: string
        pronostico:
          type: string
        proximaCita:
          type: string
        observaciones:
          type: string
      required:
        - fecha
        - motivo
...
    Movimiento:
      type: object
      properties:
        id:
          type: number
        clienteId:
          type: number
        pacienteId:
          type: ["number", "null"]
        paciente:
          type: ["string", "null"]
        fecha:
          type: string
        tipo:
          type: string
        concepto:
          type: string
        importe:
          type: number
        metodoPago:
          type: ["string", "null"]
        referencia:
          type: ["string", "null"]
        notas:
          type: ["string", "null"]
        creadoEn:
          type: string
      required:
        - id
        - clienteId
        - fecha
        - tipo
        - concepto
        - importe
        - creadoEn
    MovimientoInput:
      type: object
      properties:
        pacienteId:
          type: number
        fecha:
          type: string
          minLength: 1
        tipo:
          type: string
          enum: [Cargo, Pago, Ajuste]
        concepto:
          type: string
          minLength: 1
        importe:
          type: number
          minimum: 0
        metodoPago:
          type: string
        referencia:
          type: string
        notas:
          type: string
      required:
        - fecha
        - tipo
        - concepto
        - importe