# lib/api-zod/src/generated/types/movimientoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * VetControl API — veterinary clinic management system
 * OpenAPI spec version: 0.1.0
 */
import type { MovimientoInputTipo } from './movimientoInputTipo';

export interface MovimientoInput {
  pacienteId?: number;
  /** @minLength 1 */
  fecha: string;
  tipo: MovimientoInputTipo;
  /** @minLength 1 */
  concepto: string;
  /** @minimum 0 */
  importe: number;
  metodoPago?: string;
  referencia?: string;
  notas?: string;
}